Move variations table component to variations report directory.
